requirements.txt
Flask==1.1.2
google-cloud-bigquery==2.14.0
